requirements.txt
torch>=2.1.0
torchvision>=0.16.0
openai>=1.10.0
pyyaml>=6.0
numpy>=1.24.0
tqdm>=4.65.0
pytest>=7.4.0
python-dotenv>=1.0.0

